fix(global): split footer to server/client component
